+++ packages/core/src/components/async-server/async-server.step-builder.ts
@@ -22,6 +22,7 @@
 
 import type { AsyncClientMessageType, AsyncServerMessageType, IAsyncProtocol } from "../../protocols/base";
 import { BaseStepBuilder } from "../base/step-builder";
+import type { ValueOrFactory } from "../base/step.types";
 import type { AsyncServer } from "./async-server.component";
 import { AsyncServerHookBuilder } from "./async-server.hook-builder";
 import type { ExtractEventPayload, ExtractMessagePayload } from "./async-server.types";
@@ -227,7 +228,7 @@ export class AsyncServerStepBuilder<P extends IAsyncProtocol = IAsyncProtocol> e
 	sendEvent<K extends AsyncServerMessageType<P>>(
 		linkId: string,
 		eventType: K,
-		payload: ExtractEventPayload<P, K>
+		payload: ValueOrFactory<ExtractEventPayload<P, K>>
 	): void {
 		this.registerStep({
 			type: "sendEvent",
@@ -248,7 +249,7 @@ export class AsyncServerStepBuilder<P extends IAsyncProtocol = IAsyncProtocol> e
 	 * @param eventType - Event type to send
 	 * @param payload - Event payload
 	 */
-	broadcast<K extends AsyncServerMessageType<P>>(eventType: K, payload: ExtractEventPayload<P, K>): void {
+	broadcast<K extends AsyncServerMessageType<P>>(eventType: K, payload: ValueOrFactory<ExtractEventPayload<P, K>>): void {
 		this.registerStep({
 			type: "broadcast",
 			description: `Broadcast event ${eventType}`,